+++ src/frontend/components/SøknadsSteg/VelgBarn/LeggTilBarn/useLeggTilBarn.tsx
@@ -117,7 +117,6 @@ export const useLeggTilBarn = (): {
         if (!kanSendeSkjema()) {
             return false;
         }
-        // TODO: Bedre dato-løsning
         settSøknad({
             ...søknad,
             barnRegistrertManuelt: søknad.barnRegistrertManuelt.concat([
